Fix bad class name for ModelListType
